--- src/integration/java/README.md
@@ -1,1 +1,0 @@
-## Integration test
